requirements/tests.txt
pytest==3.0.7
pytest-selenium==1.9.1
pytest-variables==1.6.0
pyotp==2.2.4
pytest-xdist==1.18.1